DB setup is now a setup step.
